requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
pwinput==1.0.3
Pygments==2.17.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
typer==0.12.1
typing_extensions==4.11.0
urllib3==2.2.1
